Update Services all tests working - next step Controllers

+++ src/pballtest/Services/Game/DeleteGameAsyncTests.cs
@@ -14,7 +14,7 @@ public partial class GameServiceTests : BaseServiceTests
 
         if (ContactService != null)
         {
-            RegisterModel registerModel = await FillRegisterModel();
+            RegisterModel registerModel = await FillRegisterModelAsync();
 
             var actionRegisterRes = await ContactService.RegisterAsync(registerModel);
             Contact? contact = await DoOKTestReturnContactAsync(actionRegisterRes);
@@ -25,7 +25,28 @@ public partial class GameServiceTests : BaseServiceTests
                 Assert.True(contact.ContactID > 0);
             }
 
-            RegisterModel registerModel2 = await FillRegisterModel();
+            LoginModel loginModel = new LoginModel()
+            {
+                LoginEmail = registerModel.LoginEmail,
+                Password = registerModel.Password,
+            };
+
+            var actionLoginRes = await ContactService.LoginAsync(loginModel);
+            contact = await DoOKTestReturnContactAsync(actionLoginRes);
+            Assert.NotNull(contact);
+
+            if (contact != null)
+            {
+                if (UserService != null)
+                {
+                    UserService.User = contact;
+                }
+                Assert.True(contact.ContactID > 0);
+                Assert.Empty(contact.PasswordHash);
+                Assert.NotEmpty(contact.Token);
+            }
+
+            RegisterModel registerModel2 = await FillRegisterModelAsync();
 
             registerModel2.LoginEmail = "P" + registerModel.LoginEmail;
             registerModel2.FirstName = "P" + registerModel.FirstName;
@@ -41,7 +62,7 @@ public partial class GameServiceTests : BaseServiceTests
                 Assert.True(contact2.ContactID > 0);
             }
 
-            RegisterModel registerModel3 = await FillRegisterModel();
+            RegisterModel registerModel3 = await FillRegisterModelAsync();
 
             registerModel3.LoginEmail = "w" + registerModel.LoginEmail;
             registerModel3.FirstName = "w" + registerModel.FirstName;
@@ -57,7 +78,7 @@ public partial class GameServiceTests : BaseServiceTests
                 Assert.True(contact3.ContactID > 0);
             }
 
-            RegisterModel registerModel4 = await FillRegisterModel();
+            RegisterModel registerModel4 = await FillRegisterModelAsync();
 
             registerModel4.LoginEmail = "r" + registerModel.LoginEmail;
             registerModel4.FirstName = "r" + registerModel.FirstName;
@@ -202,17 +223,55 @@ public partial class GameServiceTests : BaseServiceTests
     {
         Assert.True(await _GameServiceSetupAsync(culture));
 
-        if (LoggedInService != null)
-        {
-            LoggedInService.LoggedInContactInfo.LoggedInContact = null;
-        }
+        bool? boolRet = await ClearAllContactsFromDBAsync();
+        Assert.True(boolRet);
 
-        if (GameService != null)
+        if (ContactService != null)
         {
-            var actionDeleteRes = await GameService.DeleteGameAsync(1);
-            bool? boolRet = await DoBadRequestGameTestAsync(PBallRes.YouDoNotHaveAuthorization, actionDeleteRes);
-            Assert.NotNull(boolRet);
-            Assert.True(boolRet);
+            RegisterModel registerModel = await FillRegisterModelAsync();
+
+            var actionRegisterRes = await ContactService.RegisterAsync(registerModel);
+            Contact? contact = await DoOKTestReturnContactAsync(actionRegisterRes);
+            Assert.NotNull(contact);
+
+            if (contact != null)
+            {
+                Assert.True(contact.ContactID > 0);
+            }
+
+            LoginModel loginModel = new LoginModel()
+            {
+                LoginEmail = registerModel.LoginEmail,
+                Password = registerModel.Password,
+            };
+
+            var actionLoginRes = await ContactService.LoginAsync(loginModel);
+            contact = await DoOKTestReturnContactAsync(actionLoginRes);
+            Assert.NotNull(contact);
+
+            if (contact != null)
+            {
+                if (UserService != null)
+                {
+                    UserService.User = contact;
+                }
+                Assert.True(contact.ContactID > 0);
+                Assert.Empty(contact.PasswordHash);
+                Assert.NotEmpty(contact.Token);
+            }
+
+            if (UserService != null)
+            {
+                UserService.User = null;
+            }
+
+            if (GameService != null)
+            {
+                var actionDeleteRes = await GameService.DeleteGameAsync(1);
+                boolRet = await DoBadRequestGameTestAsync(PBallRes.YouDoNotHaveAuthorization, actionDeleteRes);
+                Assert.NotNull(boolRet);
+                Assert.True(boolRet);
+            }
         }
     }
     [Theory]
@@ -222,21 +281,59 @@ public partial class GameServiceTests : BaseServiceTests
     {
         Assert.True(await _GameServiceSetupAsync(culture));
 
-        if (GameService != null)
+        bool? boolRet = await ClearAllContactsFromDBAsync();
+        Assert.True(boolRet);
+
+        if (ContactService != null)
         {
-            int GameID = 0;
+            RegisterModel registerModel = await FillRegisterModelAsync();
 
-            var actionDeleteRes = await GameService.DeleteGameAsync(GameID);
-            bool? boolRet = await DoBadRequestGameTestAsync(string.Format(PBallRes._IsRequired, "GameID"), actionDeleteRes);
-            Assert.NotNull(boolRet);
-            Assert.True(boolRet);
+            var actionRegisterRes = await ContactService.RegisterAsync(registerModel);
+            Contact? contact = await DoOKTestReturnContactAsync(actionRegisterRes);
+            Assert.NotNull(contact);
 
-            GameID = 10000;
+            if (contact != null)
+            {
+                Assert.True(contact.ContactID > 0);
+            }
 
-            actionDeleteRes = await GameService.DeleteGameAsync(GameID);
-            boolRet = await DoBadRequestGameTestAsync(string.Format(PBallRes.CouldNotFind_With_Equal_, "Game", "GameID", GameID.ToString()), actionDeleteRes);
-            Assert.NotNull(boolRet);
-            Assert.True(boolRet);
+            LoginModel loginModel = new LoginModel()
+            {
+                LoginEmail = registerModel.LoginEmail,
+                Password = registerModel.Password,
+            };
+
+            var actionLoginRes = await ContactService.LoginAsync(loginModel);
+            contact = await DoOKTestReturnContactAsync(actionLoginRes);
+            Assert.NotNull(contact);
+
+            if (contact != null)
+            {
+                if (UserService != null)
+                {
+                    UserService.User = contact;
+                }
+                Assert.True(contact.ContactID > 0);
+                Assert.Empty(contact.PasswordHash);
+                Assert.NotEmpty(contact.Token);
+            }
+
+            if (GameService != null)
+            {
+                int GameID = 0;
+
+                var actionDeleteRes = await GameService.DeleteGameAsync(GameID);
+                boolRet = await DoBadRequestGameTestAsync(string.Format(PBallRes._IsRequired, "GameID"), actionDeleteRes);
+                Assert.NotNull(boolRet);
+                Assert.True(boolRet);
+
+                GameID = 10000;
+
+                actionDeleteRes = await GameService.DeleteGameAsync(GameID);
+                boolRet = await DoBadRequestGameTestAsync(string.Format(PBallRes.CouldNotFind_With_Equal_, "Game", "GameID", GameID.ToString()), actionDeleteRes);
+                Assert.NotNull(boolRet);
+                Assert.True(boolRet);
+            }
         }
     }
 }
